+++ src/opengl/texture/arrayTexture.cpp
@@ -5,7 +5,7 @@ void ArrayTexture2D::uploadData(GLenum target, GLubyte *data ){
     bind(0);
     glTexImage2D(target,
                  0,  // level, 0 = base, no minimap,
-                 internal_format, // internalformat
+                 static_cast<GLint>(internal_format), // internalformat
                  width,  // width
                  height,  // height
                  0,  // border, always 0 in OpenGL ES
@@ -24,7 +24,7 @@ void ArrayTexture2D::uploadData(GLubyte *data ){
     for (int i=0; i<6; i++) {
         glTexImage2D(GL_TEXTURE_CUBE_MAP_POSITIVE_X + i,
                      0,  // level, 0 = base, no minimap,
-                     internal_format, // internalformat
+                     static_cast<GLint>(internal_format), // internalformat
                      width,  // width
                      height,  // height
                      0,  // border, always 0 in OpenGL ES
@@ -43,7 +43,7 @@ void ArrayTexture2D::uploadData(GLubyte **data ){
     for (int i=0; i<6; i++) {
         glTexImage2D(GL_TEXTURE_CUBE_MAP_POSITIVE_X + i,
                      0,  // level, 0 = base, no minimap,
-                     internal_format, // internalformat
+                     static_cast<GLint>(internal_format), // internalformat
                      width,  // width
                      height,  // height
                      0,  // border, always 0 in OpenGL ES
@@ -57,11 +57,11 @@ void ArrayTexture2D::uploadData(GLubyte **data ){
 }
 
 void ArrayTexture2D::setDefaultParameters(){
-    glTexParameteri (target, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-    glTexParameteri (target, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-    glTexParameteri (target, GL_TEXTURE_WRAP_R, GL_CLAMP_TO_EDGE);
-    glTexParameteri (target, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
-    glTexParameteri (target, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
+    glTexParameteri (target, GL_TEXTURE_MAG_FILTER, static_cast<GLint>(GL_LINEAR));
+    glTexParameteri (target, GL_TEXTURE_MIN_FILTER, static_cast<GLint>(GL_LINEAR));
+    glTexParameteri (target, GL_TEXTURE_WRAP_R, static_cast<GLint>(GL_CLAMP_TO_EDGE));
+    glTexParameteri (target, GL_TEXTURE_WRAP_S, static_cast<GLint>(GL_CLAMP_TO_EDGE));
+    glTexParameteri (target, GL_TEXTURE_WRAP_T, static_cast<GLint>(GL_CLAMP_TO_EDGE));
 }
